# public/openapi.yaml
openapi: 3.1.0
info:
  title: Tilli Dashboard API
  version: 1.0.0
  description: |
    API for accessing Tilli Dashboard data and translations.
    This API allows third-party integrations to retrieve assessment scores,
    teacher surveys, and localized content for displaying the dashboard in different languages.
    
    The API integrates with Appwrite backend for data storage and retrieval.
  contact:
    name: Tilli Dashboard API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-production-domain.com
    description: Production server (update with your actual domain when deploying)
paths:
  /api/scores:
    get:
      summary: Get assessment scores
      description: |
        Retrieves assessment scores from the Appwrite database. 
        Can return all scores or filtered scores based on query parameters.
        
        If all filter parameters (school, grade, assessment, section, zone) are provided,
        returns filtered results. Otherwise, returns all scores.
      operationId: getScores
      tags:
        - Assessments
      parameters:
        - name: school
          in: query
          required: false
          description: School identifier
          schema:
            type: string
            example: "azmi_mufti_boys_p1"
        - name: grade
          in: query
          required: false
          description: Grade level
          schema:
            type: string
            example: "grade1"
        - name: assessment
          in: query
          required: false
          description: Assessment type
          schema:
            type: string
            enum: [child, teacher_report, parent]
            example: "child"
        - name: section
          in: query
          required: false
          description: Section identifier
          schema:
            type: string
            example: "a"
        - name: zone
          in: query
          required: false
          description: Zone identifier
          schema:
            type: string
            enum: [irbid, north_amman, south_amman, zarqa]
            example: "irbid"
      responses:
        "200":
          description: Successfully retrieved scores
          content:
            application/json:
              schema:
                type: object
                properties:
                  scores:
                    type: array
                    items:
                      type: object
                      properties:
                        $id:
                          type: string
                          description: Document ID
                        school:
                          type: string
                        grade:
                          type: string
                        assessment:
                          type: string
                          enum: [child, teacher_report, parent]
                        total_students:
                          type: number
                        testType:
                          type: string
                          enum: [PRE, POST]
                        overall_level_distribution:
                          type: object
                          properties:
                            beginner:
                              type: number
                            growth:
                              type: number
                            expert:
                              type: number
                        category_level_distributions:
                          type: object
                          properties:
                            self_awareness:
                              $ref: "#/components/schemas/LevelDistribution"
                            social_management:
                              $ref: "#/components/schemas/LevelDistribution"
                            social_awareness:
                              $ref: "#/components/schemas/LevelDistribution"
                            relationship_skills:
                              $ref: "#/components/schemas/LevelDistribution"
                            responsible_decision_making:
                              $ref: "#/components/schemas/LevelDistribution"
                            metacognition:
                              $ref: "#/components/schemas/LevelDistribution"
                            empathy:
                              $ref: "#/components/schemas/LevelDistribution"
                            critical_thinking:
                              $ref: "#/components/schemas/LevelDistribution"
              example:
                scores:
                  - $id: "123456"
                    school: "azmi_mufti_boys_p1"
                    grade: "grade1"
                    assessment: "child"
                    total_students: 25
                    testType: "PRE"
                    overall_level_distribution:
                      beginner: 8
                      growth: 12
                      expert: 5
                    category_level_distributions:
                      self_awareness:
                        beginner: 5
                        growth: 15
                        expert: 5
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
...
components:
  schemas:
    LevelDistribution:
      type: object
      properties:
        beginner:
          type: number
          description: Number of students in beginner level
        growth:
          type: number
          description: Number of students in growth level
        expert:
          type: number
          description: Number of students in expert level
      example:
        beginner: 5
        growth: 15
        expert: 5
...